+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog for v0.x
 
+## [Unreleased]
+
+### Added
+
+- Support for VIFE `E001 1101` (Standard conform data content)
+- Test for LAN-WMBUS-G2-LDS/LDP
+
 ## v0.2.0
 
 Open-sourced